chore: Delete old forms and rename new ones (#1045)
