--- src/components/forms/pool_actions/InvestForm.vue
@@ -1,734 +1,0 @@
-<template>
-  <BalForm ref="investForm" @on-submit="submit">
-    <FormTypeToggle
-      v-model="investType"
-      :form-types="formTypes"
-      :missing-prices="missingPrices"
-      :has-zero-balance="hasZeroBalance"
-      :loading="loading"
-    />
-
-    <template v-if="isProportional">
-      <div class="p-4 border-t dark:border-gray-900">
-        <div class="border dark:border-gray-900 rounded-lg shadow-inner p-2">
-          <div
-            class="flex items-center justify-between mb-3 text-sm text-gray-600"
-          >
-            <span v-text="$t('amountToInvest')" />
-            <span>{{ propPercentage }}%</span>
-          </div>
-          <div class="flex items-end">
-            <span
-              class="mr-2 text-lg font-medium w-1/2 break-words leading-none"
-              :title="total"
-            >
-              <template v-if="missingPrices">-</template>
-              <span v-else>{{ total }}</span>
-            </span>
-            <BalRangeInput
-              class="w-1/2"
-              v-model="range"
-              :max="1000"
-              :interval="1"
-              :min="0"
-              tooltip="none"
-              :disabled="loading"
-            />
-          </div>
-        </div>
-      </div>
-
-      <div
-        :class="[
-          'px-4 py-3 bg-gray-50 dark:bg-gray-850 border-b dark:border-gray-900',
-          hasZeroBalance ? '' : 'border-t'
-        ]"
-      >
-        <div
-          v-for="(token, i) in pool.tokenAddresses"
-          :key="token"
-          class="py-3 last:mb-0"
-        >
-          <div class="flex items-center justify-between">
-            <div class="flex items-center w-1/2">
-              <BalAsset :address="token" class="mr-2" />
-              <div class="flex flex-col w-3/4 leading-none">
-                <span
-                  class="break-words"
-                  :title="
-                    `${fNum(amounts[i], 'token')} ${
-                      pool.onchain.tokens[token].symbol
-                    }`
-                  "
-                >
-                  {{ fNum(amounts[i], 'token') }}
-                  {{ pool.onchain.tokens[token].symbol }}
-                </span>
-                <span
-                  class="text-xs text-gray-400 break-words"
-                  :title="`${formatBalance(i)} balance`"
-                >
-                  {{ $t('balance') }}: {{ formatBalance(i) }}
-                </span>
-              </div>
-            </div>
-            <div class="flex flex-col w-1/2 leading-none text-right pl-2">
-              <span class="break-words" :title="fNum(amountUSD(i), 'usd')">
-                {{ amountUSD(i) === '0' ? '-' : fNum(amountUSD(i), 'usd') }}
-              </span>
-              <span v-if="!isStableLikePool" class="text-xs text-gray-400">
-                {{ fNum(tokenWeights[i], 'percent_lg') }}
-              </span>
-            </div>
-          </div>
-        </div>
-      </div>
-    </template>
-
-    <div
-      v-else
-      :class="[
-        'px-4 pt-6 border-b dark:border-gray-900',
-        hasZeroBalance ? '' : 'border-t'
-      ]"
-    >
-      <TokenInput
-        v-for="(tokenAddress, i) in tokenAddresses"
-        :key="tokenAddress"
-        v-model:amount="amounts[i]"
-        v-model:address="tokenAddresses[i]"
-        v-model:isValid="validInputs[i]"
-        :weight="isStableLikePool ? 0 : Number(tokenWeights[i])"
-        :name="tokenAddress"
-        class="mb-4"
-        fixedToken
-      />
-
-      <div v-if="isWethPool" class="flex items-center mb-4">
-        <router-link
-          :to="{
-            name: 'trade',
-            params: {
-              assetIn: appNetworkConfig.nativeAsset.address,
-              assetOut: appNetworkConfig.addresses.weth
-            }
-          }"
-          class="text-xs text-gray-500 underline"
-        >
-          {{ $t('wrapInstruction', [nativeAsset, `W${nativeAsset}`]) }}
-        </router-link>
-        <BalTooltip>
-          <template v-slot:activator>
-            <BalIcon name="info" size="xs" class="text-gray-400 ml-2" />
-          </template>
-          <div v-html="$t('ethBufferInstruction')" />
-        </BalTooltip>
-      </div>
-      <div v-if="isWstETHPool" class="flex items-center mb-4">
-        <router-link
-          :to="{
-            name: 'trade',
-            params: {
-              assetIn: appNetworkConfig.addresses.stETH,
-              assetOut: appNetworkConfig.addresses.wstETH
-            }
-          }"
-          class="text-xs text-gray-500 underline"
-        >
-          {{
-            $t('wrapInstruction', [
-              symbolFor(appNetworkConfig.addresses.stETH),
-              symbolFor(appNetworkConfig.addresses.wstETH)
-            ])
-          }}
-        </router-link>
-        <BalTooltip>
-          <template v-slot:activator>
-            <BalIcon name="info" size="xs" class="text-gray-400 ml-2" />
-          </template>
-          <div v-html="$t('wrapStEthTooltip')" />
-        </BalTooltip>
-      </div>
-    </div>
-
-    <div class="p-4">
-      <BalBtn
-        v-if="!isWalletReady"
-        :label="$t('connectWallet')"
-        block
-        @click.prevent="toggleWalletSelectModal"
-      />
-      <template v-else>
-        <div
-          :class="['flex items-center text-sm mb-4', priceImpactClasses]"
-          @click.prevent
-        >
-          <span>
-            {{ $t('priceImpact') }}: {{ fNum(priceImpact, 'percent') }}
-          </span>
-          <BalTooltip>
-            <template v-slot:activator>
-              <BalIcon
-                v-if="priceImpact >= 0.01"
-                name="alert-triangle"
-                size="xs"
-                class="ml-1"
-              />
-              <BalIcon
-                v-else
-                name="info"
-                size="xs"
-                class="text-gray-400 -mb-px ml-2"
-              />
-            </template>
-            <div v-html="$t('customAmountsTip')" />
-          </BalTooltip>
-        </div>
-        <BalBtn
-          v-if="requireApproval"
-          :label="`${$t('approve')} ${symbolFor(requiredApprovals[0])}`"
-          :loading="approving"
-          :loading-label="$t('approving')"
-          :disabled="!hasAmounts || !hasValidInputs"
-          block
-          @click.prevent="approveNextAllowance"
-        />
-        <template v-else>
-          <BalCheckbox
-            v-if="priceImpact >= 0.01"
-            v-model="highPiAccepted"
-            :rules="[isRequired($t('priceImpactCheckbox'))]"
-            name="highPiAccepted"
-            class="text-gray-500 mb-12"
-            size="sm"
-            :label="$t('priceImpactAccept', [$t('depositing')])"
-          />
-          <BalBtn
-            type="submit"
-            :loading-label="$t('confirming')"
-            color="gradient"
-            :disabled="!hasAmounts || !hasValidInputs || isMismatchedNetwork"
-            :loading="loading"
-            block
-            @click="trackGoal(Goals.ClickInvest)"
-          >
-            <span>{{ $t('invest') }}</span>
-            <span v-if="!(missingPrices || total.length > 15)"
-              >&nbsp;{{ total }}</span
-            >
-          </BalBtn>
-        </template>
-      </template>
-    </div>
-  </BalForm>
-</template>
-
-<script lang="ts">
-import {
-  defineComponent,
-  computed,
-  watch,
-  onMounted,
-  reactive,
-  toRefs,
-  PropType,
-  toRef,
-  ref
-} from 'vue';
-import { FormRef } from '@/types';
-import {
-  isPositive,
-  isLessThanOrEqualTo,
-  isRequired
-} from '@/lib/utils/validations';
-import { useI18n } from 'vue-i18n';
-import { formatUnits } from '@ethersproject/units';
-import isEqual from 'lodash/isEqual';
-
-import useTokenApprovals from '@/composables/pools/useTokenApprovals';
-import useNumbers from '@/composables/useNumbers';
-import useSlippage from '@/composables/useSlippage';
-
-import PoolExchange from '@/services/pool/exchange/exchange.service';
-import PoolCalculator from '@/services/pool/calculator/calculator.sevice';
-import { getPoolWeights } from '@/services/pool/pool.helper';
-import { bnum } from '@/lib/utils';
-import FormTypeToggle from './shared/FormTypeToggle.vue';
-import { FullPool } from '@/services/balancer/subgraph/types';
-import useFathom from '@/composables/useFathom';
-
-import { TOKENS } from '@/constants/tokens';
-import useWeb3 from '@/services/web3/useWeb3';
-import useTokens from '@/composables/useTokens';
-import { TransactionReceipt } from '@ethersproject/abstract-provider';
-import useEthers from '@/composables/useEthers';
-import useTransactions from '@/composables/useTransactions';
-import { usePool } from '@/composables/usePool';
-import TokenInput from '@/components/inputs/TokenInput/TokenInput.vue';
-
-export enum FormTypes {
-  proportional = 'proportional',
-  custom = 'custom'
-}
-
-type DataProps = {
-  investForm: FormRef;
-  investType: FormTypes;
-  loading: boolean;
-  tokenAddresses: string[];
-  amounts: string[];
-  propMax: string[];
-  validInputs: boolean[];
-  propToken: number;
-  range: number;
-  highPiAccepted: boolean;
-};
-
-export default defineComponent({
-  name: 'InvestForm',
-
-  components: {
-    FormTypeToggle,
-    TokenInput
-  },
-
-  emits: ['success'],
-
-  props: {
-    pool: { type: Object as PropType<FullPool>, required: true },
-    missingPrices: { type: Boolean, default: false }
-  },
-
-  setup(props: { pool: FullPool }, { emit }) {
-    const data = reactive<DataProps>({
-      investForm: {} as FormRef,
-      investType: FormTypes.proportional,
-      loading: false,
-      tokenAddresses: props.pool.tokenAddresses,
-      amounts: [],
-      propMax: [],
-      validInputs: [],
-      propToken: 0,
-      range: 1000,
-      highPiAccepted: false
-    });
-
-    // COMPOSABLES
-    const {
-      isWalletReady,
-      isMismatchedNetwork,
-      account,
-      toggleWalletSelectModal,
-      getProvider,
-      appNetworkConfig
-    } = useWeb3();
-    const { fNum, toFiat } = useNumbers();
-    const { t } = useI18n();
-    const { minusSlippage } = useSlippage();
-    const { tokens, balances: allBalances } = useTokens();
-    const { trackGoal, Goals } = useFathom();
-    const { txListener } = useEthers();
-    const { addTransaction } = useTransactions();
-    const {
-      isStableLikePool,
-      isWethPool,
-      isWstETHPool,
-      managedPoolWithTradingHalted
-    } = usePool(toRef(props, 'pool'));
-
-    // SERVICES
-    const poolExchange = new PoolExchange(toRef(props, 'pool'));
-
-    const poolCalculator = new PoolCalculator(
-      toRef(props, 'pool'),
-      tokens,
-      allBalances,
-      'join',
-      ref(false)
-    );
-
-    // COMPUTED
-    const tokenWeights = computed(() =>
-      Object.values(props.pool.onchain.tokens).map(t => t.weight)
-    );
-
-    const hasAmounts = computed(() => {
-      const amountSum = fullAmounts.value
-        .map(amount => parseFloat(amount))
-        .reduce((a, b) => a + b, 0);
-      return amountSum > 0;
-    });
-
-    const hasValidInputs = computed(() => {
-      return data.validInputs.every(validInput => validInput === true);
-    });
-
-    const balances = computed(() => {
-      return props.pool.tokenAddresses.map(
-        token => allBalances.value[token] || '0'
-      );
-    });
-
-    const hasZeroBalance = computed(() => {
-      return balances.value.map(b => Number(b)).includes(0);
-    });
-
-    const total = computed(() => {
-      const total = props.pool.tokenAddresses
-        .map((_, i) => amountUSD(i))
-        .reduce(
-          (a, b) =>
-            bnum(a)
-              .plus(b)
-              .toNumber(),
-          0
-        );
-
-      if (total < 0) return fNum(0, 'usd');
-      return fNum(total, 'usd');
-    });
-
-    const requireApproval = computed(() => {
-      if (!hasAmounts.value) return false;
-      return requiredApprovals.value.length > 0;
-    });
-
-    const isProportional = computed(() => {
-      return data.investType === FormTypes.proportional;
-    });
-
-    const propPercentage = computed(() => {
-      const currentAmount = fullAmounts.value[data.propToken];
-      const maxAmount = tokenBalance(data.propToken);
-
-      if (currentAmount === '0') return 0;
-      return Math.ceil((Number(currentAmount) / Number(maxAmount)) * 100);
-    });
-
-    const fullAmounts = computed(() => {
-      return props.pool.tokenAddresses.map((_, i) => {
-        return data.amounts[i] || '0';
-      });
-    });
-
-    const propMaxUSD = computed(() => {
-      const total = props.pool.tokenAddresses
-        .map((token, i) => toFiat(data.propMax[i], token))
-        .reduce(
-          (a, b) =>
-            bnum(a)
-              .plus(b)
-              .toNumber(),
-          0
-        );
-
-      return fNum(total, 'usd');
-    });
-
-    const balanceMaxUSD = computed(() => {
-      const total = props.pool.tokenAddresses
-        .map((token, i) => toFiat(balances.value[i], token))
-        .reduce(
-          (a, b) =>
-            bnum(a)
-              .plus(b)
-              .toNumber(),
-          0
-        );
-
-      return fNum(total, 'usd');
-    });
-
-    const priceImpact = computed(() => {
-      if (isProportional.value || !hasAmounts.value) return 0;
-      return poolCalculator.priceImpact(fullAmounts.value).toNumber() || 0;
-    });
-
-    const priceImpactClasses = computed(() => {
-      return {
-        'text-red-500 font-medium': priceImpact.value >= 0.01,
-        'text-gray-500 font-normal': priceImpact.value < 0.01
-      };
-    });
-
-    const bptOut = computed(() => {
-      let bptOut = poolCalculator
-        .exactTokensInForBPTOut(fullAmounts.value)
-        .toString();
-
-      return formatUnits(bptOut, props.pool.onchain.decimals);
-    });
-
-    const minBptOut = computed(() => {
-      return minusSlippage(bptOut.value, props.pool.onchain.decimals);
-    });
-
-    const nativeAsset = computed(() => appNetworkConfig.nativeAsset.symbol);
-
-    const formTypes = computed(() => {
-      let validTypes = [
-        {
-          label: t('noPriceImpact'),
-          max: propMaxUSD.value,
-          value: FormTypes.proportional,
-          tooltip: t('noPriceImpactTip')
-        }
-      ];
-
-      // Managed pools with trading halted only allow proportional joins/exits
-      if (!managedPoolWithTradingHalted.value) {
-        validTypes.push({
-          label: t('customAmounts'),
-          max: balanceMaxUSD.value,
-          value: FormTypes.custom,
-          tooltip: t('customAmountsTip')
-        });
-      }
-
-      return validTypes;
-    });
-
-    const {
-      approving,
-      requiredApprovals,
-      approveNextAllowance
-    } = useTokenApprovals(props.pool.tokenAddresses, fullAmounts);
-
-    // METHODS
-    function tokenBalance(index: number): string {
-      return balances.value[index] || '0';
-    }
-
-    function tokenDecimals(index) {
-      return tokens.value[props.pool.tokenAddresses[index]].decimals;
-    }
-
-    function amountUSD(index) {
-      const amount = fullAmounts.value[index] || '0';
-      return toFiat(amount, props.pool.tokenAddresses[index]);
-    }
-
-    function formatBalance(index) {
-      return fNum(tokenBalance(index), 'token');
-    }
-
-    function amountRules(index) {
-      return isWalletReady.value
-        ? [
-            isPositive(),
-            isLessThanOrEqualTo(
-              Number(tokenBalance(index)),
-              t('exceedsBalance')
-            )
-          ]
-        : [isPositive()];
-    }
-
-    function symbolFor(token: string) {
-      return tokens.value[token]?.symbol || '';
-    }
-
-    async function setPropMax() {
-      const { send, fixedToken } = poolCalculator.propMax();
-      data.propMax = [...send];
-      data.propToken = fixedToken;
-    }
-
-    function resetSlider() {
-      data.amounts = [...data.propMax];
-      data.range = 1000;
-    }
-
-    function setPropAmountsFor(range) {
-      const fractionBasisPoints = (range / 1000) * 10000;
-      const amount = bnum(balances.value[data.propToken])
-        .times(fractionBasisPoints)
-        .div(10000)
-        .toFixed(tokenDecimals(data.propToken));
-
-      const { send } = poolCalculator.propAmountsGiven(
-        amount,
-        data.propToken,
-        'send'
-      );
-      data.amounts = send;
-    }
-
-    // Legacy function for sense check against JS calculation of BPT out
-    // Left here so numbers can be debugged in conosle
-    // Talk to Fernando to see if still needed
-    async function calcMinBptOut(): Promise<void> {
-      let { bptOut: queryBptOut } = await poolExchange.queryJoin(
-        getProvider(),
-        account.value,
-        fullAmounts.value,
-        props.pool.tokenAddresses,
-        minBptOut.value
-      );
-      queryBptOut = formatUnits(
-        queryBptOut.toString(),
-        props.pool.onchain.decimals
-      );
-      console.log(queryBptOut, 'bptOut (queryJoin)');
-      console.log(minBptOut.value, 'bptOut (JS) minusSlippage');
-      console.log(
-        minusSlippage(queryBptOut, props.pool.onchain.decimals),
-        'bptOut (queryJoin) minusSlippage'
-      );
-    }
-
-    async function submit(): Promise<void> {
-      if (!data.investForm.validate()) return;
-      try {
-        data.loading = true;
-        await calcMinBptOut();
-        const _bptOut = managedPoolWithTradingHalted.value
-          ? bptOut.value
-          : minBptOut.value;
-
-        const tx = await poolExchange.join(
-          getProvider(),
-          account.value,
-          fullAmounts.value,
-          props.pool.tokenAddresses,
-          _bptOut
-        );
-        console.log('Receipt', tx);
-
-        addTransaction({
-          id: tx.hash,
-          type: 'tx',
-          action: 'invest',
-          summary: t('transactionSummary.investInPool', [
-            total.value,
-            getPoolWeights(props.pool)
-          ]),
-          details: {
-            total,
-            pool: props.pool
-          }
-        });
-
-        txListener(tx, {
-          onTxConfirmed: async (tx: TransactionReceipt) => {
-            emit('success', tx);
-            data.amounts = [];
-            data.loading = false;
-            setPropMax();
-            resetSlider();
-          },
-          onTxFailed: () => {
-            data.loading = false;
-          }
-        });
-      } catch (error) {
-        console.error(error);
-        data.loading = false;
-      }
-    }
-
-    watch(
-      () => props.pool.onchain.tokens,
-      (newTokens, oldTokens) => {
-        const tokensChanged = !isEqual(newTokens, oldTokens);
-        if (tokensChanged) {
-          setPropMax();
-          if (isProportional.value) setPropAmountsFor(data.range);
-        }
-      }
-    );
-
-    watch(balances, (newBalances, oldBalances) => {
-      const balancesChanged = !isEqual(newBalances, oldBalances);
-      if (balancesChanged) {
-        setPropMax();
-        if (isProportional.value) setPropAmountsFor(data.range);
-      }
-    });
-
-    watch(
-      () => data.investType,
-      newType => {
-        if (newType === FormTypes.proportional) {
-          setPropMax();
-          resetSlider();
-        }
-      }
-    );
-
-    watch(
-      () => data.range,
-      newVal => {
-        setPropAmountsFor(newVal);
-      }
-    );
-
-    watch(isWalletReady, isAuth => {
-      if (!isAuth) {
-        data.amounts = [];
-        data.propMax = [];
-      }
-    });
-
-    watch(account, () => {
-      if (hasZeroBalance.value) {
-        data.investType = FormTypes.custom;
-      } else {
-        setPropMax();
-        resetSlider();
-      }
-    });
-
-    onMounted(() => {
-      if (hasZeroBalance.value) {
-        data.investType = FormTypes.custom;
-      } else {
-        setPropMax();
-        resetSlider();
-      }
-    });
-
-    return {
-      // data
-      ...toRefs(data),
-      Goals,
-      nativeAsset,
-      TOKENS,
-      // computed
-      tokens,
-      appNetworkConfig,
-      hasValidInputs,
-      hasAmounts,
-      approving,
-      requireApproval,
-      requiredApprovals,
-      tokenWeights,
-      tokenBalance,
-      amountRules,
-      total,
-      isWalletReady,
-      isMismatchedNetwork,
-      toggleWalletSelectModal,
-      formatBalance,
-      isProportional,
-      propPercentage,
-      priceImpact,
-      priceImpactClasses,
-      amountUSD,
-      formTypes,
-      isRequired,
-      hasZeroBalance,
-      isWethPool,
-      isWstETHPool,
-      isStableLikePool,
-      // methods
-      submit,
-      approveNextAllowance,
-      fNum,
-      trackGoal,
-      symbolFor,
-      tokenDecimals
-    };
-  }
-});
-</script>
